+++ CHANGELOG.md
@@ -1,7 +1,7 @@
-### NEXT_VERSION_TYPE=MINOR
-### NEXT_VERSION_DESCRIPTION_BEGIN
+## [2.16.0]() (24-12-2019)
+
 * Add the ability to configure dashboard annotation with Graphite datasource. See `AnnotationDemo.kts`
-### NEXT_VERSION_DESCRIPTION_END
+
 ## [2.15.0]() (17-12-2019)
 
 * Add an `alignTo` attribute to the `summarize` function